Allow for User concurrency
The web API server now handles concurrent users via multithreading. ensuring only one connection can connect at a time. This prevents race conditions, where multiple threads attempt to modify or access shared data such as writing to the consle in this case.

Also makes use of a lock guard, which manages the locking and unlocking of a thread in a scoped manner. It locks the mutex when the lock_guard object is created and automatically unlocks it when the object goes out of scope.